% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_stage.R
\name{first_stage}
\alias{first_stage}
\title{Calculates the first stage of a sorting model}
\usage{
first_stage(code_name, Z_names, X_names, dat)
}
\arguments{
\item{code_name}{Indicates (with name or column number) the vector with alternative chosen}

\item{Z_names}{Indicates (with names or column numbers) the vectors with individual data}

\item{X_names}{Indicates (with names or column numbers) the vectors with city/regional data}

\item{dat}{Dataset to be used}
}
\value{
A (maxlik) object with the estimates (maxlik) together with the indicators of the alternative chosen,
 the vectors with individual data, the vectors with city/regional data and the name of the reference alternative
}
\description{
Calculates the first stage of a sorting model
}
\details{
Optimization occurs via the maxLik package for an obtimization of a conditional logit model
}
\examples{
data <- municipality
model_output <- first_stage("mun_code", c("age","income"),
                c("lnprice","monuments"), data)
}
