% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipality}
\alias{municipality}
\title{A test dataset of individual locational choices}
\format{A data frame with 9,074 rows and 19 variables:}
\usage{
municipality
}
\description{
A test dataset of individual locational choices
}
\keyword{datasets}
