% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_stage.R
\name{second_stage}
\alias{second_stage}
\title{Title Estimates the second of a sorting model}
\usage{
second_stage(s1.results, dat, endog = NULL, instr = NULL)
}
\arguments{
\item{s1.results}{Indicates the (maxLik) object estimation results of the first stage of the sorting model}

\item{dat}{Dataset to be used}

\item{endog}{Indicates the endogenous variable to be instrumented (from the dataset in parentheses)}

\item{instr}{indicates  the intrument for the endogenous variable}
}
\value{
An estimation object
}
\description{
Title Estimates the second of a sorting model
}
\details{
This function estimates an OLS or an instrument variable.
if an instrument variable procedure is needed both the instrument and the endogenous variables should be given
}
\examples{
data <- municipality
model_output <- first_stage("mun_code", c("age","income"),
                c("lnprice","monuments"), data)
endog <- ("lnprice")
phat <- sorting_inst(model_output, "lnprice", data)

second_stage(model_output, data)
second_stage(model_output, data, "lnprice", phat[[1]])

}
