% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_inst.R
\name{sorting_inst}
\alias{sorting_inst}
\title{Calculates an instrument for an endogenous variable in a sorting model setting}
\usage{
sorting_inst(s1.results, endog, dat, n.iterations = 3, stepsize = 0.05)
}
\arguments{
\item{s1.results}{Indicates the (maxLik) object estimation results of the first stage of the sorting model}

\item{endog}{Indicates the endogenous variable to be instrumented}

\item{dat}{Dataset to be used}

\item{n.iterations}{Indicates the number of iterations}

\item{stepsize}{Indicates the contraction-mapping scaling coefficient}
}
\value{
A list containing (1) a vector of the computed instrument, and (2) the correlation between
the computed instrument and the original variable.
}
\description{
Calculates an instrument for an endogenous variable in a sorting model setting
}
\details{
Instrument is calculated using a logit model estimation, assuming market clearing conditions
given no unobserved heterogeneity between alternative choices (following Bayer et al. (2004))
}
\examples{
data <- municipality
model_output <- first_stage("mun_code", c("age","income"),
                c("lnprice","monuments"), data)
phat <- sorting_inst(model_output, "lnprice", data)
}
