% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortingmod.R
\docType{package}
\name{sortingmod}
\alias{sortingmod}
\alias{sortingmod-package}
\title{sortingmod: a package for the estimation of sorting models}
\description{
The sortingmod package provides two important functions:
first_stage and second_stage.
}
\section{sortingmod functions}{

The sortingmod functions
}

