% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_stage.R
\name{first_stage}
\alias{first_stage}
\title{Estimates the first stage of a sorting model}
\usage{
first_stage(code_name, Z_names, X_names, data, print_detail = 3)
}
\arguments{
\item{code_name}{Indicates (with name or column number) the vector with alternative chosen}

\item{Z_names}{Indicates (with names or column numbers) the vectors with individual data}

\item{X_names}{Indicates (with names or column numbers) the vectors with alternative data}

\item{data}{Dataset to be used}

\item{print_detail}{Controls the output detail of maxLik (default is set at 3)}
}
\value{
A (maxLik) object with the estimates (maxLik) together with the indicators of the alternative chosen,
 the vectors with individual data, the vectors with city/regional data and the name of the reference alternative
}
\description{
Estimates the first stage of a sorting model
}
\details{
Optimization occurs via the maxLik package for an obtimization of a conditional logit model
}
\examples{
code_name <- c("mun_code")
Z_names <- c("age","income")
X_names <- c("lnprice","nature","monuments")
data <- municipality
s1.results <- first_stage(code_name, Z_names, X_names, data, print_detail = 1)
}
