% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_stage.R
\name{second_stage}
\alias{second_stage}
\title{Title Estimates the second of a sorting model}
\usage{
second_stage(s1.results, data, endog = NULL, instr = NULL)
}
\arguments{
\item{s1.results}{Indicates the (maxLik) object estimation results of the first stage of the sorting model}

\item{endog}{Indicates the endogenous variable(s) to be instrumented (from the dataset in parentheses)}

\item{instr}{indicates the intrument(s) for the endogenous variable}

\item{dat}{Dataset to be used}
}
\value{
An estimation object
}
\description{
Title Estimates the second of a sorting model
}
\details{
This function estimates an OLS or an instrument variable.
if an instrument variable procedure is needed both the instrument and the endogenous variables should be given
}
\examples{
data <- municipality
s1.results <- first_stage(code_name = "mun_code",
                         X_names = c("lnprice","kindergardens_1km","p_mig_west","nature","monuments","cafes_1km"),
                         Z_names = c("income","double_earner_hh","hh_kids","age", "migskill"),
                         data = data)
endog <- ("lnprice")
phat <- sorting_inst(s1.results, "lnprice", data, stepsize = 0.02)

second_stage(s1.results, data)
second_stage(s1.results, data, "lnprice", phat[[1]])

}
